refactor: break down cache storage.rs into focused modules
- Split 1284-line storage.rs into organized module hierarchy
- All files under 200 lines (largest: 161 lines)
- Group by function: backend/, wal/, format, transaction
- Use directory hierarchies, no underscores
- Maintain all functionality and test coverage
